move scram to kafka package
