+++ pkg/sink/kafka/scram_client.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package security
+package kafka
 
 import (
 	"crypto/sha256"
@@ -22,21 +22,17 @@ import (
 )
 
 var (
-	// SHA256 func
-	SHA256 scram.HashGeneratorFcn = func() hash.Hash { return sha256.New() }
-	// SHA512 func
-	SHA512 scram.HashGeneratorFcn = func() hash.Hash { return sha512.New() }
+	sha256HashGenerator scram.HashGeneratorFcn = func() hash.Hash { return sha256.New() }
+	sha512HashGenerator scram.HashGeneratorFcn = func() hash.Hash { return sha512.New() }
 )
 
-// XDGSCRAMClient xdg scram client
-type XDGSCRAMClient struct {
+type xdgSCRAMClient struct {
 	*scram.Client
 	*scram.ClientConversation
 	scram.HashGeneratorFcn
 }
 
-// Begin xdg scram client Begin
-func (x *XDGSCRAMClient) Begin(userName, password, authzID string) (err error) {
+func (x *xdgSCRAMClient) Begin(userName, password, authzID string) (err error) {
 	x.Client, err = x.NewClient(userName, password, authzID)
 	if err != nil {
 		return err
@@ -45,13 +41,11 @@ func (x *XDGSCRAMClient) Begin(userName, password, authzID string) (err error) {
 	return nil
 }
 
-// Step xdg scram client Step
-func (x *XDGSCRAMClient) Step(challenge string) (response string, err error) {
+func (x *xdgSCRAMClient) Step(challenge string) (response string, err error) {
 	response, err = x.ClientConversation.Step(challenge)
 	return
 }
 
-// Done xdg scram client Done
-func (x *XDGSCRAMClient) Done() bool {
+func (x *xdgSCRAMClient) Done() bool {
 	return x.ClientConversation.Done()
 }
